+++ src/yarpmanager/gybuilder/src/int_port_model.cpp
@@ -69,11 +69,11 @@ Gdk::Point InternalPortModel::getContactPoint(void)
             GooCanvasBounds bi;
             goo_canvas_item_get_bounds(item, &bi);
             if(type == INPUTD)
-                return Gdk::Point(bi.x1+1, bi.y2-PORT_SIZE/2.0);
-            return Gdk::Point(bi.x2-1, bi.y2-PORT_SIZE/2.0);
+                return Gdk::Point((int)(bi.x1+1), (int)(bi.y2-PORT_SIZE/2.0));
+            return Gdk::Point((int)(bi.x2-1), (int)(bi.y2-PORT_SIZE/2.0));
         }
     }     
-    return Gdk::Point(-1, -1);       
+    return Gdk::Point(-1, -1);
 }
