+++ src/yarpmanager/gybuilder/src/label_model.cpp
@@ -49,7 +49,7 @@ Gdk::Point LabelModel::getContactPoint(void)
         GooCanvasBounds bi;
         goo_canvas_item_get_bounds(item, &bi);
         goo_canvas_item_get_bounds(item, &bi);
-        return Gdk::Point(bi.x2, bi.y2);
+        return Gdk::Point((int)(bi.x2), (int)(bi.y2));
     }    
     return Gdk::Point(-1, -1);
 }
